requirements.txt
django
djangorestframework
python-decouple
django-environ